Update solution from xproj to new csproj
* Update to new csproj

* Fix inline elements that use starting capturing group

Some inline elements e.g. Monospace, use a starting capturing group as part of regular expression matching, to capture the preceding character in group 1.

When replacing the match in the block being parsed with whitespace, put the contents of this capturing group back into the block as part of the replacement. Move the replacement string into the inline match to use for replacement.

This inline element parsing logic is horribly organised and inefficient, but works for now, until it can be refactored.

* Fix image output
* Fix nuget packaging
* Fix build and xunit warnings

Suppress public member XML comment warnings for now as they are very noisy